Update show_results
A prior show_results update added returned results for scrapbook, but this broke multidomain support which was not tested previously. Current testing revealed this resulting in show_results needing to be updated. This update returns a single dataframe consisting of the sum of all instabilities as well as the overall max velocity (i.e. not for each domain).